# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Memtara API
  version: 0.1.0
  summary: Zero-knowledge attribute disclosure, and the signed attestations that carry its verdicts off-platform.
  description: |
    Memtara lets a consumer prove a *predicate* about themselves — income above a
    threshold, residency valid, no sanctions match — without disclosing the
    attribute behind it. Proofs are generated on the holder's device against one
    of four Noir circuits; this API authenticates users, syncs the encrypted
    vault, orchestrates the disclosure-request lifecycle, verifies submitted
    proofs against the real compiled circuits via Barretenberg, and issues
    short-lived Ed25519-signed attestations that any JOSE-speaking relying party
    can validate offline.

    **This API never generates proofs.** Generating one requires the plaintext
    attribute as a witness, so a server that could prove would be a server that
    holds plaintext. See `ARCHITECTURE.md`. Every "verified" verdict here comes
    from running the real circuit's verifier over bytes a client produced.

    **Compliance context.** `docs/REGULATORY_MATRIX.md` maps this system, clause
    by clause, against the CBUAE *Guidance Note on Consumer Protection and the
    Responsible Adoption and Use of AI and ML by LFIs* (`CBUAE_EN_6958_VER1`).
    Predicate responses carry the clause ids each disclosure is evidence for.
  license:
    name: See repository
  contact:
    name: Memtara
...
servers:
...
security:
  - sessionToken: []
paths:
...
  /api/v1/wealth-assessments/{request_id}:
    get:
      tags: [Structured-product suitability]
      summary: The evidence pack for one assessment
      security:
        - orgApiKey: []
      parameters:
        - name: request_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      description: |
        Everything needed to reconstruct one suitability assessment, in one
        call: the terms and where they came from, every proof attempt (accepted
        *and* rejected), the verification key digest, the audit-chain excerpt,
        and the verdict with a note on how it was read.

        This is an API rather than a report generator on purpose. A bank's
        compliance function will want the evidence in their own template, in
        Arabic, joined to their own CRM; if the only route to it were a script
        written against our schema, they would be blocked on us for every one of
        those. `scripts/export_audit_evidence.py` is simply the consumer that
        renders a PDF.

        What is deliberately absent: the client's income, liquid assets, risk
        tolerance and holdings. The server never had them, and the response says
        so explicitly rather than leaving a reader to wonder whether the gap is
        an oversight.
      responses:
        "200":
          description: The evidence pack.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: No such assessment for this organisation.
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        An opaque, revocable session token from any auth flow. Opaque rather
        than a JWT precisely so it can be killed in one query — the product's
        revocation guarantees depend on it.
    orgApiKey:
      type: http
      scheme: bearer
      description: |
        A relying party's API key, presented in the same `Authorization: Bearer`
        header as a session token. Routes readable by "the org that created it
        or the user it is for" try the presented credential as both.
...
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...